kelar

+++ app/Http/Controllers/ListingController.php
@@ -44,6 +44,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created succesfully!');
@@ -51,11 +53,19 @@ class ListingController extends Controller
 
     // Show Edit Form
     public function edit(Listing $listing) {
+        // Make sure current user is authorized
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
         return view('listings.edit', ['listing' => $listing]);
     }
 
     // Update Listing Data
     public function update(Request $request, Listing $listing) {
+        // Make sure current user is authorized
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
 
         $formFields = $request->validate([
             'title' => 'required',
@@ -78,8 +88,19 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        // Make sure current user is authorized
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted succesfully');
     }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
+
 }
